Moved detach to utility pallet
