+++ pallets/fragments/Cargo.toml
@@ -74,12 +74,6 @@ default-features = false
 features = ['derive']
 version = '1.0'
 
-[dependencies.pallet-randomness-collective-flip]
-default-features = false
-git = 'https://github.com/paritytech/substrate.git'
-tag = 'monthly-2022-01'
-version = '4.0.0-dev'
-
 [dependencies]
 log = { version = "0.4.14", default-features = false }
 sp-chainblocks = { path = '../../primitives/chainblocks', default-features = false }
@@ -96,7 +90,6 @@ std = [
     'sp-io/std',
     'sp-std/std',
 	'sp-runtime/std',
-    'pallet-randomness-collective-flip/std',
     'log/std',
 	'sp-keystore',
 	'sp-core/std',
